+++ examples/demos/glost-ko-transcription-example/CHANGELOG.md
@@ -1,5 +1,21 @@
 # glost-ko-transcription-example
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - glost-transcription@0.4.6
+  - glost-ko@0.1.6
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - glost-transcription@0.4.5
+  - glost-ko@0.1.5
+
 ## 0.0.5
 
 ### Patch Changes