+++ packages/languages/ja/CHANGELOG.md
@@ -1,5 +1,19 @@
 # glost-ja
 
+## 0.4.6
+
+### Patch Changes
+
+- Updated dependencies
+  - glost-transcription@0.4.6
+
+## 0.4.5
+
+### Patch Changes
+
+- Updated dependencies
+  - glost-transcription@0.4.5
+
 ## 0.4.4
 
 ### Patch Changes